+++ src/Actions/Messages/StoreImageMessage.php
@@ -102,8 +102,7 @@ class StoreImageMessage extends NewMessageAction
             ->setType('image')
             ->setDisk($this->getThread()->getStorageDisk())
             ->setDirectory($this->getDirectory())
-            ->upload($file)
-            ->getName();
+            ->upload($file);
     }
 
     /**