+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 - All notable changes to `Messenger` will be documented in this file.
 
+---
+
+### [v0.29 (05-02-2021)]
+
+#### Changed
+- FileService now resets properties after each `upload`, `destroy` and `destroyDirectory`.
+
+#### Removed
+- `getName()` method removed from FileService. `upload()` method returns final file name now.
+
+---
+
 ### [v0.28 (04-22-2021)]
 
 #### Added
@@ -12,6 +24,8 @@
 #### Removed
 - `getProviderId()` and `getProviderClass()` methods removed from Messenger.
 
+---
+
 ### [v0.27 (04-21-2021)]
 
 #### Removed
@@ -20,6 +34,8 @@
 #### Changed
 - `instance()` method on messenger singleton renamed to `getInstance()`
 - General major refactoring.
+
+---
 
 ### [v0.26 (04-12-2021)]
 
@@ -32,6 +48,8 @@
 #### Changed
 - Renamed action class UpdateMessage to EditMessage.
 
+---
+
 ### [v0.25 (04-10-2021)]
 
 #### Added
@@ -39,6 +57,8 @@
 
 #### Changed
 - Either party in a private call can end the call.
+
+---
 
 ### [v0.24 (04-05-2021)]
 
@@ -55,6 +75,8 @@
 - New emoji picker added to the included UI.
 - Emoji converter is now a service/interface.
 
+---
+
 ### [v0.23 (03-24-2021)]
 
 #### Added
@@ -63,6 +85,8 @@
 
 #### Changed
 - All store message actions accept params as array now instead of individual params.
+
+---
 
 ### [v0.22 (03-15-2021)]
 
@@ -78,6 +102,8 @@
 #### Removed
 - `message_documents.download` toggle in config removed.
 
+---
+
 ### [v0.21 (03-11-2021)]
 
 #### Added
@@ -86,11 +112,15 @@
 #### Changed
 - threads, messages, participants and message_edits tables use precision 6 for timestamps now.
 
+---
+
 ### [v0.20 (03-09-2021)]
 
 #### Changed
 - Allow more mime types on uploads, frontend assets updated with this as well.
 - Misc bug fixes along with image service not resizing gif/svg/webp.
+
+---
 
 ### [v0.19 (02-28-2021)]
 
@@ -98,11 +128,15 @@
 - New configs to set upload limit sizes.
 - New commands to temporarily shutdown calling system and end all active calls, as well as put the system back online.
 
+---
+
 ### [v0.18 (02-20-2021)]
 
 #### Changed
 - More file moves/renames.
 - Set broadcast/video driver on demand.
+
+---
 
 ### [v0.17 (02-19-2021)]
 
@@ -113,11 +147,15 @@
 - New Helpers class and support directory. Moved files around.
 - Added intermediate modal to confirm joining call once page loads.
 
+---
+
 ### [v0.16 (02-17-2021)]
 
 #### Changed
 - Bugfix on join call skipping generating participant resource on response.
 - Broadcast broker resets private/presence each time to method called.
+
+---
 
 ### [v0.15 (02-14-2021)]
 
@@ -128,6 +166,8 @@
 - Edit message action will store the edit history immediately after update message.
 - Put a primary key back onto messenger model and table messengers.
 
+---
+
 ### [v0.14 (02-09-2021)]
 
 #### Added
@@ -137,6 +177,8 @@
 - Exceptions thrown throughout out package.
 - Moved some authorization logic done in controllers or model into actions.
 
+---
+
 ### [v0.13 (02-07-2021)]
 
 #### Added
@@ -144,25 +186,35 @@
 - Route to view edit history.
 - Config option to disable both edit message and viewing edit history.
 
+---
+
 ### [v0.12 (02-05-2021)]
 
 #### Added
 - Edit message feature.
+
+---
 
 ### [v0.11 (02-03-2021)]
 
 #### Changed
 - `teardown_complete` added to calls table used to avoid duplicate tear downs. Added a short cache lockout upon ending a call to avoid race conditions with automated EndCallIfEmpty job.
 
+---
+
 ### [v0.10 (02-03-2021)]
 
 #### Changed
 - To avoid conflicts with channel names across apps, our channels are now prefixed with `messenger.` All impacted test and frontend assets have been updated to reflect this change.
 
+---
+
 ### [v0.9 (02-01-2021)]
 
 #### Added
 - Test.
+
+---
 
 ### [v0.1]